wip wip wip need to propagate register allocation
