+++ tests/filecheck/backend/riscv/canonicalize.mlir
@@ -30,6 +30,9 @@ builtin.module {
   // Don't optimise out unused immediates
   "test.op"(%zero, %0, %1, %2, %3) : (!riscv.reg<zero>, !riscv.reg<>, !riscv.reg<>, !riscv.reg<>, !riscv.reg<>) -> ()
 
+  %load_zero_zero = riscv.li 0 : () -> !riscv.reg<zero>
+  "test.op"(%load_zero_zero) : (!riscv.reg<zero>) -> ()
+
   %add_immediate_zero_reg = riscv.addi %zero, 1 : (!riscv.reg<zero>) -> !riscv.reg<a0>
   "test.op"(%add_immediate_zero_reg) : (!riscv.reg<a0>) -> ()
 
@@ -130,11 +133,14 @@ builtin.module {
 // CHECK-NEXT:   "test.op"(%f0, %fo1, %fo2, %f0, %fo4, %fo5) : (!riscv.freg<fa0>, !riscv.freg<fa2>, !riscv.freg<>, !riscv.freg<fa0>, !riscv.freg<fa2>, !riscv.freg<>) -> ()
 
 // CHECK-NEXT:   %zero = riscv.get_register : () -> !riscv.reg<zero>
-// CHECK-NEXT:   %0 = riscv.li 0 : () -> !riscv.reg<zero>
+// CHECK-NEXT:   %0 = riscv.li 0 : () -> !riscv.reg<>
 // CHECK-NEXT:   %1 = riscv.li 1 : () -> !riscv.reg<>
 // CHECK-NEXT:   %2 = riscv.li 2 : () -> !riscv.reg<>
 // CHECK-NEXT:   %3 = riscv.li 3 : () -> !riscv.reg<>
-// CHECK-NEXT:   "test.op"(%zero, %0, %1, %2, %3) : (!riscv.reg<zero>, !riscv.reg<zero>, !riscv.reg<>, !riscv.reg<>, !riscv.reg<>) -> ()
+// CHECK-NEXT:   "test.op"(%zero, %0, %1, %2, %3) : (!riscv.reg<zero>, !riscv.reg<>, !riscv.reg<>, !riscv.reg<>, !riscv.reg<>) -> ()
+
+// CHECK-NEXT:   %load_zero_zero = riscv.get_register : () -> !riscv.reg<zero>
+// CHECK-NEXT:   "test.op"(%load_zero_zero) : (!riscv.reg<zero>) -> ()
 
 // CHECK-NEXT:   %add_immediate_zero_reg = riscv.li 1 : () -> !riscv.reg<a0>
 // CHECK-NEXT:   "test.op"(%add_immediate_zero_reg) : (!riscv.reg<a0>) -> ()
