+++ tests/filecheck/backend/riscv/riscv_register_allocation.mlir
@@ -4,6 +4,7 @@
 riscv_func.func @external() -> ()
 
 riscv_func.func @main() {
+  %zero = riscv.li 0 : () -> !riscv.reg<>
   %0 = riscv.li 6 : () -> !riscv.reg<>
   %1 = riscv.li 5 : () -> !riscv.reg<s0>
   %2 = riscv.fcvt.s.w %0 : (!riscv.reg<>) -> !riscv.freg<>
@@ -25,6 +26,7 @@ riscv_func.func @main() {
 //   CHECK-LIVENESS-BLOCK-NAIVE:       builtin.module {
 //   CHECK-LIVENESS-BLOCK-NAIVE-NEXT:    riscv_func.func @external() -> ()
 //   CHECK-LIVENESS-BLOCK-NAIVE-NEXT:    riscv_func.func @main() {
+//   CHECK-LIVENESS-BLOCK-NAIVE-NEXT:      %{{.+}} = riscv.li 0 : () -> !riscv.reg<zero>
 //   CHECK-LIVENESS-BLOCK-NAIVE-NEXT:      %{{\d+}} = riscv.li 6 : () -> !riscv.reg<t2>
 //   CHECK-LIVENESS-BLOCK-NAIVE-NEXT:      %{{\d+}} = riscv.li 5 : () -> !riscv.reg<s0>
 //   CHECK-LIVENESS-BLOCK-NAIVE-NEXT:      %{{\d+}} = riscv.fcvt.s.w %{{\d+}} : (!riscv.reg<t2>) -> !riscv.freg<ft0>
@@ -45,6 +47,7 @@ riscv_func.func @main() {
 //   CHECK-LIVENESS-BLOCK-NAIVE-J:       builtin.module {
 //   CHECK-LIVENESS-BLOCK-NAIVE-J-NEXT:    riscv_func.func @external() -> ()
 //   CHECK-LIVENESS-BLOCK-NAIVE-J-NEXT:    riscv_func.func @main() {
+//   CHECK-LIVENESS-BLOCK-NAIVE-J-NEXT:      %{{.+}} = riscv.li 0 : () -> !riscv.reg<zero>
 //   CHECK-LIVENESS-BLOCK-NAIVE-J-NEXT:      %{{\d+}} = riscv.li 6 : () -> !riscv.reg<j2>
 //   CHECK-LIVENESS-BLOCK-NAIVE-J-NEXT:      %{{\d+}} = riscv.li 5 : () -> !riscv.reg<s0>
 //   CHECK-LIVENESS-BLOCK-NAIVE-J-NEXT:      %{{\d+}} = riscv.fcvt.s.w %{{\d+}} : (!riscv.reg<j2>) -> !riscv.freg<j3>
