+++ tests/filecheck/projects/riscv-backend-paper/ddot_regalloc.mlir
@@ -34,7 +34,7 @@ riscv.assembly_section ".text" {
 // CHECK-NEXT:          %Y_moved = riscv.mv %Y : (!riscv.reg<a1>) -> !riscv.reg<t1>
 // CHECK-NEXT:          %Z_moved = riscv.mv %Z : (!riscv.reg<a2>) -> !riscv.reg<t0>
 // CHECK-NEXT:          %init = riscv.fld %Z_moved, 0 : (!riscv.reg<t0>) -> !riscv.freg<ft0>
-// CHECK-NEXT:          %lb = riscv.li 0 : () -> !riscv.reg<t4>
+// CHECK-NEXT:          %lb = riscv.li 0 : () -> !riscv.reg<zero>
 // CHECK-NEXT:          %ub = riscv.li 1024 : () -> !riscv.reg<t5>
 // CHECK-NEXT:          %c8 = riscv.li 8 : () -> !riscv.reg<t6>
 // CHECK-NEXT:          %res = riscv_scf.for %i : !riscv.reg<t3> = %lb to %ub step %c8 iter_args(%acc_in = %init) -> (!riscv.freg<ft0>) {
